requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
blinker==1.9.0
boto3==1.35.92
botocore==1.35.92
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
faiss-cpu==1.9.0.post1
filelock==3.16.1
fonttools==4.55.3
fpdf==1.7.2
fpdf2==2.8.2
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
idna==3.10
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.13
langchain-aws==0.2.11
langchain-community==0.3.13
langchain-core==0.3.28
langchain-experimental==0.3.4
langchain-mistralai==0.2.4
langchain-text-splitters==0.3.4
langsmith==0.2.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
mdurl==0.1.2
mistralai==1.3.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.20.1
networkx==3.4.2
numpy==2.2.1
openpyxl==3.1.5
orjson==3.10.13
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
protobuf==5.29.2
pyarrow==18.1.0
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
s3transfer==0.10.4
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.15.0
sentence-transformers==3.3.1
setuptools==75.6.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.36
streamlit==1.41.1
streamlit-chat==0.1.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.47.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
watchdog==6.0.0
yarl==1.18.3
